requirements.txt
APScheduler==3.10.4
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.7
Django==5.0.3
django-apscheduler==0.6.2
django-cors-headers==4.3.1
django-filter==24.2
django-rest-knox==4.2.0
djangorestframework==3.15.1
dynaconf==3.2.5
hvac==2.2.0
idna==3.7
Jinja2==3.1.4
knox==0.1.14
loguru==0.7.2
MarkupSafe==2.1.5
pycparser==2.22
pytz==2024.1
requests==2.32.2
shortuuid==1.0.13
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzlocal==5.2
urllib3==2.2.1
validators==0.28.3
